requirements/base.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-nested-routers==0.95.0
django-filter==25.2
python-dotenv==1.1.0
